requirements.txt
pyspark==3.5.3
kafka-python==2.0.2
pandas==2.0.3
numpy==1.24.3
